input before name to match others
