+++ src/packages/documents/document-types/components/index.ts
@@ -1,1 +1,1 @@
-import './document-type-input/document-type-input.element.js';
+import './input-document-type/input-document-type.element.js';
