+++ src/packages/documents/document-types/workspace/views/structure/document-type-workspace-view-structure.element.ts
@@ -1,5 +1,5 @@
 import { UmbDocumentTypeWorkspaceContext } from '../../document-type-workspace.context.js';
-import type { UmbDocumentTypeInputElement } from '../../../components/document-type-input/document-type-input.element.js';
+import type { UmbInputDocumentTypeElement } from '../../../components/input-document-type/input-document-type.element.js';
 import { css, html, customElement, state } from '@umbraco-cms/backoffice/external/lit';
 import { UUITextStyles } from '@umbraco-cms/backoffice/external/uui';
 import type { UUIToggleElement } from '@umbraco-cms/backoffice/external/uui';
@@ -62,10 +62,10 @@ export class UmbDocumentTypeWorkspaceViewStructureElement
 					</div>
 					<div slot="editor">
 						<!-- TODO: maybe we want to somehow display the hierarchy, but not necessary in the same way as old backoffice? -->
-						<umb-document-type-input
+						<umb-input-document-type
 							.selectedIds=${this._allowedContentTypeIDs}
 							@change="${(e: CustomEvent) => {
-								const sortedContentTypesList = (e.target as UmbDocumentTypeInputElement).selectedIds.map(
+								const sortedContentTypesList = (e.target as UmbInputDocumentTypeElement).selectedIds.map(
 									(id, index) => ({
 										id: id,
 										sortOrder: index,
@@ -73,7 +73,7 @@ export class UmbDocumentTypeWorkspaceViewStructureElement
 								);
 								this.#workspaceContext?.setAllowedContentTypes(sortedContentTypesList);
 							}}">
-						</umb-document-type-input>
+						</umb-input-document-type>
 					</div>
 				</umb-workspace-property-layout>
 			</uui-box>
