+++ src/packages/documents/documents/components/input-document/input-document.test.ts
@@ -1,15 +1,15 @@
 import { expect, fixture, html } from '@open-wc/testing';
-import { UmbDocumentInputElement } from './document-input.element.js';
+import { UmbInputDocumentElement } from './input-document.element.js';
 import { defaultA11yConfig } from '@umbraco-cms/internal/test-utils';
-describe('UmbDocumentInputElement', () => {
-	let element: UmbDocumentInputElement;
+describe('UmbInputDocumentElement', () => {
+	let element: UmbInputDocumentElement;
 
 	beforeEach(async () => {
-		element = await fixture(html` <umb-document-input></umb-document-input> `);
+		element = await fixture(html` <umb-input-document></umb-input-document> `);
 	});
 
 	it('is defined with its own instance', () => {
-		expect(element).to.be.instanceOf(UmbDocumentInputElement);
+		expect(element).to.be.instanceOf(UmbInputDocumentElement);
 	});
 
 	it('passes the a11y audit', async () => {
